backend/requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
Werkzeug
flask-cors
flask-jwt-extended
python-dotenv
flasgger